warning fixes, changed floats to doubles so you can go deeper with detail, and disabled warnings enabled by /Wall

+++ src/generator.cpp
@@ -7,15 +7,15 @@ Generator::Generator(float screenWidth, float screenHeight) {
   width = screenWidth;
   height = screenHeight;
 
-  initRealStart = -2.0f;
-  initRealEnd = 2.0f;
-  initImagStart = -2.0f;
-  initImagEnd = 2.0f;
+  initRealStart = -2.0;
+  initRealEnd = 2.0;
+  initImagStart = -2.0;
+  initImagEnd = 2.0;
 
-  realStart = -2.0f;
-  realEnd = 2.0f;
-  imagStart = -2.0f;
-  imagEnd = 2.0f;
+  realStart = -2.0;
+  realEnd = 2.0;
+  imagStart = -2.0;
+  imagEnd = 2.0;
 
   itcap = 256;
 
@@ -26,22 +26,22 @@ Generator::Generator(float screenWidth, float screenHeight) {
 
 }
 
-float* Generator::getComplexNumber(int x, int y, float realStart, float realEnd, float imagStart, float imagEnd, int width, int height) {
+double* Generator::getComplexNumber(int x, int y, double realStart, double realEnd, double imagStart, double imagEnd, int width, int height) {
 
-  float c[2] = {(realStart + (((float)x/width) * (realEnd - realStart))), (imagStart + (((float)y/height) * (imagEnd -imagStart)))};
+  double c[2] = {(realStart + (((double)x/width) * (realEnd - realStart))), (imagStart + (((double)y/height) * (imagEnd -imagStart)))};
 
   return c;
 } 
 
-void Generator::generate(float*(*pointFunction)(float, float, int)) {
+void Generator::generate(double*(*pointFunction)(double, double, int)) {
   for (int pixel=0; pixel<(sizeof(pixels)/sizeof(pixels[0])); pixel++) {
-    float* c = getComplexNumber((pixel % width), (int)(pixel / height), realStart, realEnd, imagStart, imagEnd, width, height);
+    double* c = getComplexNumber((pixel % width), (int)(pixel / height), realStart, realEnd, imagStart, imagEnd, width, height);
 
 
-    float* it = pointFunction(c[0], c[1], itcap);
+    double* it = pointFunction(c[0], c[1], itcap);
 
     if ((int)it[0] == itcap) {
-      pixels[pixel] = 0.0f;
+      pixels[pixel] = 0.0;
     }
     else {
       // pixels[pixel] = (it[0] + 1 - log(log(sqrt(it[1] * it[1] + it[2] * it[2]))/log(2))) / itcap;
